+++ release-notes.md
@@ -1,5 +1,9 @@
 Release Notes
 -------------
+#### v1.0.3rc
+- Add statistical prediction
+- Refactoring (move code to mathstat)
+
 #### v1.0.2 (2017-04-20)
 - Fix #20: Preserve user config after reinstall
 - Fix #16: Remove changing file permissions